Export vpnstatus and vpnconfig packages
