+++ internal/splitrt/splitrt_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/T-Systems-MMS/oc-daemon/internal/addrmon"
 	"github.com/T-Systems-MMS/oc-daemon/internal/devmon"
 	"github.com/T-Systems-MMS/oc-daemon/internal/dnsproxy"
-	"github.com/T-Systems-MMS/oc-daemon/internal/vpnconfig"
+	"github.com/T-Systems-MMS/oc-daemon/pkg/vpnconfig"
 	"github.com/vishvananda/netlink"
 )
